+++ docs/changelog.rst
@@ -22,6 +22,7 @@ Fixes
 * :class:`aspecd.utils.ToDictMixin`: Added superclass call to preserve mro in dependent subclasses
 * Tasks properly handle non-dataset results from recipe
 * :class:`aspecd.plotting.MultiPlotter` sets axis labels when units are empty
+* :class:`aspecd.processing.Normalisation` removes unit from last axis
 
 
 Version 0.6.2